Refactor: extract policy parsing out of Entitlement initialization
